refactor: remove permission for delete and update others movies

+++ app/Http/Controllers/MovieController.php
@@ -72,37 +72,48 @@ class MovieController extends Controller
 
 	public function update(UpdateMovieRequest $request, Movie $movie): JsonResponse
 	{
-		$movie->title = ['en' => $request->title_en, 'ka' => $request->title_ka];
-		$movie->director = ['en' => $request->director_en, 'ka' => $request->director_ka];
-		$movie->description = ['en' => $request->description_en, 'ka' => $request->description_ka];
-		$movie->year = $request->year;
-		$movie->budget = $request->budget;
-		$movie->slug = Str::slug($request->title_en);
-		if ($request->hasFile('image'))
+		if (auth()->user()->getAuthIdentifier() === $movie->user_id)
 		{
-			File::delete(public_path('images') . $movie->image);
-			$file = $request->file('image');
-			$filename = $file->getClientOriginalName();
-			$file->move('images/', $filename);
-			$movie->image = 'images/' . $filename;
+			$movie->title = ['en' => $request->title_en, 'ka' => $request->title_ka];
+			$movie->director = ['en' => $request->director_en, 'ka' => $request->director_ka];
+			$movie->description = ['en' => $request->description_en, 'ka' => $request->description_ka];
+			$movie->year = $request->year;
+			$movie->budget = $request->budget;
+			$movie->slug = Str::slug($request->title_en);
+
+			if ($request->hasFile('image'))
+			{
+				File::delete(public_path('images') . $movie->image);
+				$file = $request->file('image');
+				$filename = $file->getClientOriginalName();
+				$file->move('images/', $filename);
+				$movie->image = 'images/' . $filename;
+			}
+
+			$movie->update();
+			GenreMovie::where('movie_id', $movie->id)->delete();
+			$genres = $request->genres;
+			foreach ($genres as $genre)
+			{
+				$currentGenreId = Genre::where('name', $genre)->first()->id;
+				$genreMovie = new GenreMovie();
+				$genreMovie->movie_id = $movie->id;
+				$genreMovie->genre_id = $currentGenreId;
+				$genreMovie->save();
+			}
+			return response()->json(['message' => 'Movie updated successfully!', 200]);
 		}
-		$movie->update();
-		GenreMovie::where('movie_id', $movie->id)->delete();
-		$genres = $request->genres;
-		foreach ($genres as $genre)
-		{
-			$currentGenreId = Genre::where('name', $genre)->first()->id;
-			$genreMovie = new GenreMovie();
-			$genreMovie->movie_id = $movie->id;
-			$genreMovie->genre_id = $currentGenreId;
-			$genreMovie->save();
-		}
-		return response()->json('Quote updated successfully!');
+
+		return response()->json(['error' => 'You dont have a permission to update this movie', 404]);
 	}
 
 	public function destroy(Movie $movie): JsonResponse
 	{
-		$movie->delete();
-		return response()->json('Movie successfully deleted');
+		if (auth()->user()->getAuthIdentifier() === $movie->user_id)
+		{
+			$movie->delete();
+			return response()->json(['message' => 'Movie successfully deleted', 200]);
+		}
+		return response()->json(['error' => 'You dont have a permission to delete this movie', 404]);
 	}
 }
